Refactor api reg server to use regprocessor
